// src/types.ts
export type Team = {
  id: string;
  name: string;
  primaryColor: string;
  secondaryColor: string;
  city: string;
  state: string;
  mascot: string;
  logo: string;
  alumni: string;
  seed: number | null;
  region: string | null;
  espn_id: string | number | null;
  alumni_summary: string;
  switchlogo?: boolean;
};